+++ RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+## [0.0.3-b] - 2020-11-17
+
+- NBomber.FSharp.Http rename operation 'create' -> 'execute'
+- update all deps including FSharp.Core
+- update to NBomber 2.0
+
 ## [0.0.2] - 2020-11-17
 
 - add connection pool builder